refactor(lessonsTable): 将课程排布相关函数分离
